Clean up performance test
The spirit of the test remains the same - test the throughput of a
simple webserver, which has to do some amount of computation, within
an enclave - but the implementation has been simplified. The
documentation and scripts have also been cleaned up to simplify the
setup and execution of the test.

During testing, I was able to achieve almost 500 requests per second
for both two and four vCPUs. This is mainly due to the fact that the
ramp-up time is set to 20 seconds (the full test runs take 24
seconds). As I increase the amount of busy-looping per request, I
start to see more errors (and it's not immediately obvious what
constitutes an error). I'm leaving the test and the result as is and
will follow up with a better benchmark implementation.